Fix AI integration and Add AI chat

+++ src/AIModule/AIUrl.service.ts
@@ -6,14 +6,18 @@ import { lastValueFrom } from "rxjs";
 import { createReadStream } from "fs";
 import { CreateAssetResponseDto } from "./Dtos/CreateAssetResponse.dto";
 import { join } from "path";
+import { IAIUrlService } from "./IAIUrl.service";
+import * as stream from 'stream';
+import { ChatResponseDto } from "./Dtos/ChatResponse.dto";
+import { plainToInstance } from "class-transformer";
 
 @Injectable({ scope: Scope.REQUEST })
-export class AIUrlService {
+export class AIUrlService implements IAIUrlService{
 
 
     constructor(
         private readonly configService: ConfigService,
-        private readonly httpService: HttpService
+        private readonly httpService: HttpService,
     ) {
 
     }
@@ -41,6 +45,7 @@ export class AIUrlService {
 
     async DeleteKnowledgeBase(knowledgeBaseId: string, raiseErrorIfFailed: boolean = true): Promise<void> {
         try {
+            if(!knowledgeBaseId) return;
             await lastValueFrom(
                 this.httpService.delete(`${this.configService.getOrThrow<string>("AIBASEURL")}/knowledge_bases/${knowledgeBaseId}`)
             );
@@ -52,11 +57,25 @@ export class AIUrlService {
     }
 
 
-    async AddAsset(knowledgeBaseId: string, filePath: string): Promise<CreateAssetResponseDto> {
+    AddAsset(knowledgeBaseId: string, fileBuffer:Buffer,fileName:string): Promise<CreateAssetResponseDto> 
+    AddAsset(knowledgeBaseId: string, filePath: string): Promise<CreateAssetResponseDto>
+    async AddAsset(knowledgeBaseId: string, file?:Buffer | string,fileName?:string): Promise<CreateAssetResponseDto> {
         const FormData = require('form-data');
 
         const form = new FormData();
-        form.append('file', createReadStream(join(__dirname, "..", "..", "files", filePath)));
+        if(Buffer.isBuffer(file))
+        {
+            const readable = new stream.PassThrough();
+            readable.end(file);
+
+            form.append('file',readable,{
+                filename: fileName.substring(0,fileName.lastIndexOf('.'))+ ".pdf",
+                contentType: 'application/pdf',
+              });
+        }else
+        {
+            form.append('file',createReadStream(join(__dirname, "..", "..", "files",file)));
+        }
 
         try {
             const response = await lastValueFrom(
@@ -99,6 +118,7 @@ export class AIUrlService {
 
     async DeleteAsset(knowledgeBaseId: string, assetId: string, raiseErrorIfFailed: boolean = true) {
         try {
+            if(!knowledgeBaseId || !assetId) return;
             await lastValueFrom(
                 this.httpService.delete(`${this.configService.getOrThrow<string>("AIBASEURL")}/assets/${knowledgeBaseId}/asset/${assetId}`)
             );
@@ -109,4 +129,26 @@ export class AIUrlService {
         }
     }
 
+    async Chat(knowledgeBaseId: string, query: string): Promise<ChatResponseDto> {
+        try {
+            const response = await lastValueFrom(
+                this.httpService.post<ChatResponseDto>(`${this.configService.getOrThrow<string>("AIBASEURL")}/nlp/knowledge-bases/${knowledgeBaseId}/chat`,
+                   {
+                    "query": query,
+                    "history": [],
+                    "use_rag": true,
+                    "use_hybrid": true,
+                    "limit": 5,
+                    "use_query_rewriting": true
+                   }
+                ),
+            );
+
+            return plainToInstance(ChatResponseDto, response.data, { excludeExtraneousValues: true })
+        }
+        catch(err){
+            console.log(err)
+            throw new InternalServerErrorException("Error has happened try again later")
+        }
+    }
 }
